requirements.txt
Django==1.10.2
djangorestframework==3.4.7
gunicorn==19.6.0
psycopg2==2.6.2
